day5 no progress
we have moved some of the stuff over to win32_handmade.h to de-clutter
the code a bit.
namely the global #defines etc
