+++ code/win32_handmade.cpp
@@ -4,35 +4,7 @@
 
 #include <windows.h>
 #include <stdint.h>
-
-#define internal static
-#define local_persist static
-#define global_variable static
-
-typedef uint8_t uint8;
-typedef uint16_t uint16;
-typedef uint32_t uint32;
-typedef uint64_t uint64;
-
-typedef int8_t int8;
-typedef int16_t int16;
-typedef int32_t int32;
-typedef int64_t  int64;
-
-struct win32_offscreen_buffer {
-    BITMAPINFO Info; // FIXED(smzb): No more global variables, they exist in structs now
-    void *Memory; 
-    int Width;
-    int Height;
-    int Pitch;
-    int BytesPerPixel;
-};
-
-struct win32_window_dimensions
-{
-    int Width;
-    int Height;
-};
+#include "win32_handmade.h"
 
 win32_window_dimensions Win32GetWindowDimensions(HWND Window) {
     win32_window_dimensions Result;
@@ -42,9 +14,6 @@ win32_window_dimensions Win32GetWindowDimensions(HWND Window) {
     Result.Height = ClientRect.bottom - ClientRect.top;
     return(Result);
 }
-
-global_variable bool Running; // TODO(smzb): This should not be a global in future.
-global_variable win32_offscreen_buffer GlobalBackBuffer;
 
 internal void RenderGradient(win32_offscreen_buffer Buffer, int XOffset, int YOffset)
 {
